day 12: turning it into a db

+++ mouredev/011.py
@@ -29,6 +29,19 @@ session.headers.update({"Authorization": f"Bearer {TOKEN}"})
 def get_user():
     return session.get("https://api.github.com/user").json()["login"]
 
+def load_product(filename: str):
+    product_name = input("Nombre del producto: ")
+    q_sold = int(input("Cantidad vendida: "))
+    price= int(input("Precio: "))
+
+    f = open(f"{filename}.txt", "a")
+
+    f.write(f"{product_name}, {q_sold}, {price} \n")
+
+
+    f.close()
+
+
 def create_file(filename: str):
     f = open(f"{filename}.txt", "w")
     name = input("Ingrese su nombre: ")
@@ -39,10 +52,13 @@ def create_file(filename: str):
     f.write(f"- Tu edad: {age}\n")
     f.write(f"- Lenguaje de programación favorito: {pr_lang}\n")
 
+    f.close()
+
 
 def read_file(filename: str):
     f = open(f"{filename}.txt","r")
     print(f.read())
+    f.close()
 
 def delete_file(filename: str):
     try: 
@@ -51,11 +67,67 @@ def delete_file(filename: str):
 
     except FileNotFoundError: print(f"File '{filename}.txt' not found.")
 
+
+def describe(filename: str):
+    product = input("Ingrese el nombre del producto: ")
+
+    f = open(f"{filename}.txt", "r")
+
+    for i in f.readlines():
+        if product in i.split(",")[0]:
+            print(i)
+            return True
+    print(f"product `{product}` was not found")
+    f.close()
+    return False
+
+def update_product(filename: str):
+    read_file(filename)
+    product = input("Ingrese el nombre del producto a modificar: ")
+
+    f = open(f"{filename}.txt","r+")
+    for index, i in enumerate(f.readlines()):
+        print(index, i)
+        #if product in i.split(",")[0]:
+        #    print(i)
+        #    return True
+    #print(f"product `{product}` was not found")
+    f.close()
+
+menu = """
+1) Cargar productos
+2) Ver productos
+3) Ver producto especifico
+4) Modificar productos
+5) Ver totales de venta
+
+9) Ver el menu
+0) Salir
+"""
 if __name__ == "__main__":
     username = get_user()
-    create_file(username)
-    read_file(username)
-    delete_file(username)
+    print(menu)
+    while True:
+        option = int(input("Que desea hacer? "))
+        if option == 1:
+            load_product(username)
+        elif option == 2:
+            read_file(username)
+        elif option == 3:
+            describe(username)
+        elif option == 4:
+            update_product(username)
+        elif option == 9:
+            print(menu)
+        elif option == 0:
+            delete_file(username)
+            exit()
+
+
+
+
+    #create_file(username)
+    #delete_file(username)
 
 
     #    curl -L \
